+++ integra-portico-persistence/src/main/java/xeredi/argo/model/comun/vo/ArchivoInfoVO.java
@@ -18,7 +18,7 @@ public final class ArchivoInfoVO {
     private String nombre;
 
     /** The filesize. */
-    private Integer tamanio;
+    private Long tamanio;
 
     /** The sentido. */
     private ArchivoSentido sentido;
